T061: Extract report generator into report.js
setup.js was 1846 lines — extracted generateReport + helpers
(escHtml, getModuleDescription, getModuleSource, collectModules,
EVENT_ORDER/TITLES/BADGES) into report.js (620 lines).
setup.js is now 1261 lines (32% reduction). All 82 tests pass.
