+++ TODO.md
@@ -84,8 +84,13 @@ Modular hook runner system for Claude Code. One runner per event, modules in fol
 - [x] T059: Update README with test/uninstall commands and commit-msg-check module
 - [x] T060: Add --help command and bump version to 1.1.0
 
+## Refactor & New Features
+- [x] T061: Extract generateReport + helpers into report.js (setup.js 1846→1261 lines, report.js 620 lines)
+- [ ] T062: Add no-hardcoded-paths PreToolUse module (blocks Write/Edit with absolute paths)
+- [ ] T063: Add --upgrade command (self-updater from GitHub)
+
 ## Status
-- 60 tasks completed, 0 pending
+- 60 tasks completed, 3 pending
 - Version: 1.1.0
 - 82 tests passing across 5 test files (16 runner + 7 wizard + 13 async + 36 module + 10 sync)
 - CI: GitHub Actions runs all tests on push/PR — badge in README
